refactor(intro): split slides into self-contained StatefulWidgets
